requirements.txt
agno>=0.5.0
anthropic>=0.8.0
openai>=1.0.0
rich>=13.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
chromadb>=0.4.0
pypdf>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0